Paper on sensor design
Group built a low cost active sensor. At the end of the brief summery is
a link. Describes in depth on how the sensor was designed, the
components used, and how they programmed. Dylan, or anyone else
interested in the programming side of this project should really read
this one.